requirements.txt
Django==2.2
django-cors-headers==2.5.2
django-rest-framework==0.1.0
djangorestframework==3.9.2
pytz==2018.9
sqlparse==0.3.0
whitenoise==4.1.2
